[select] Fix position flicker when no default value is set
