+++ packages/react/src/select/arrow/SelectArrow.tsx
@@ -29,19 +29,19 @@ export const SelectArrow = React.forwardRef(function SelectArrow(
   const { className, render, ...elementProps } = componentProps;
 
   const { store } = useSelectRootContext();
-  const { arrowRef, side, align, arrowUncentered, arrowStyles } = useSelectPositionerContext();
+  const { side, align, arrowRef, arrowStyles, alignItemWithTriggerActive } =
+    useSelectPositionerContext();
 
   const open = useSelector(store, selectors.open);
-  const alignItemWithTriggerActive = useSelector(store, selectors.alignItemWithTriggerActive);
 
   const state: SelectArrow.State = React.useMemo(
     () => ({
       open,
       side,
       align,
-      uncentered: arrowUncentered,
+      uncentered: align === 'center',
     }),
-    [open, side, align, arrowUncentered],
+    [open, side, align],
   );
 
   const element = useRenderElement('div', componentProps, {
