+++ packages/react/src/select/positioner/SelectPositioner.tsx
@@ -4,12 +4,10 @@ import { useSelectRootContext, useSelectFloatingContext } from '../root/SelectRo
 import { CompositeList } from '../../composite/list/CompositeList';
 import type { BaseUIComponentProps } from '../../utils/types';
 import { popupStateMapping } from '../../utils/popupStateMapping';
-import { useSelectPositioner } from './useSelectPositioner';
-import type { Align, Side } from '../../utils/useAnchorPositioning';
+import { useAnchorPositioning, type Align, type Side } from '../../utils/useAnchorPositioning';
 import { SelectPositionerContext } from './SelectPositionerContext';
 import { InternalBackdrop } from '../../utils/InternalBackdrop';
 import { inertValue } from '../../utils/inertValue';
-import { useOnFirstRender } from '../../utils/useOnFirstRender';
 import { useModernLayoutEffect } from '../../utils/useModernLayoutEffect';
 import { useRenderElement } from '../../utils/useRenderElement';
 import { DROPDOWN_COLLISION_AVOIDANCE } from '../../utils/constants';
@@ -17,6 +15,7 @@ import { clearPositionerStyles } from '../popup/utils';
 import { useEventCallback } from '../../utils/useEventCallback';
 import { useSelector } from '../../utils/store';
 import { selectors } from '../store';
+import { useScrollLock } from '../../utils';
 
 /**
  * Positions the select menu popup.
@@ -57,31 +56,15 @@ export const SelectPositioner = React.forwardRef(function SelectPositioner(
   const value = useSelector(store, selectors.value);
   const touchModality = useSelector(store, selectors.touchModality);
   const positionerElement = useSelector(store, selectors.positionerElement);
-  const controlledItemAnchor = useSelector(store, selectors.controlledItemAnchor);
-  const alignItemWithTriggerActive = useSelector(store, selectors.alignItemWithTriggerActive);
+  const triggerElement = useSelector(store, selectors.triggerElement);
 
-  useOnFirstRender(() => {
-    const controlledItemAnchorV = alignItemWithTrigger;
-    const alignItemWithTriggerActiveV = mounted && controlledItemAnchor && !touchModality;
-    store.state.controlledItemAnchor = controlledItemAnchorV;
-    store.state.alignItemWithTriggerActive = alignItemWithTriggerActiveV;
-  });
+  const [controlledAlignItemWithTrigger, setControlledAlignItemWithTrigger] =
+    React.useState(alignItemWithTrigger);
+  const alignItemWithTriggerActive = mounted && controlledAlignItemWithTrigger && !touchModality;
 
-  useModernLayoutEffect(() => {
-    return store.subscribe((state) => {
-      const alignItemWithTriggerActiveV =
-        selectors.mounted(state) &&
-        selectors.controlledItemAnchor(state) &&
-        !selectors.touchModality(state);
-      store.set('alignItemWithTriggerActive', alignItemWithTriggerActiveV);
-    });
-  }, [store]);
-
-  useModernLayoutEffect(() => {
-    if (!mounted && controlledItemAnchor !== alignItemWithTrigger) {
-      store.set('controlledItemAnchor', alignItemWithTrigger);
-    }
-  }, [store, controlledItemAnchor, alignItemWithTrigger, mounted]);
+  if (!mounted && controlledAlignItemWithTrigger !== alignItemWithTrigger) {
+    setControlledAlignItemWithTrigger(alignItemWithTrigger);
+  }
 
   useModernLayoutEffect(() => {
     if (!alignItemWithTrigger || !mounted) {
@@ -94,11 +77,16 @@ export const SelectPositioner = React.forwardRef(function SelectPositioner(
     }
   }, [store, mounted, alignItemWithTrigger]);
 
-  React.useImperativeHandle(alignItemWithTriggerActiveRef, () =>
-    selectors.alignItemWithTriggerActive(store.state),
-  );
+  React.useImperativeHandle(alignItemWithTriggerActiveRef, () => alignItemWithTriggerActive);
 
-  const positioner = useSelectPositioner({
+  useScrollLock({
+    enabled: (alignItemWithTriggerActive || modal) && open,
+    mounted,
+    open,
+    referenceElement: triggerElement,
+  });
+
+  const positioning = useAnchorPositioning({
     anchor,
     floatingRootContext,
     positionMethod,
@@ -111,19 +99,41 @@ export const SelectPositioner = React.forwardRef(function SelectPositioner(
     collisionBoundary,
     collisionPadding,
     sticky,
-    trackAnchor,
+    trackAnchor: trackAnchor ?? !alignItemWithTriggerActive,
     collisionAvoidance,
     keepMounted: true,
   });
 
+  const positionerStyles: React.CSSProperties = React.useMemo(
+    () => (alignItemWithTriggerActive ? { position: 'fixed' } : positioning.positionerStyles),
+    [alignItemWithTriggerActive, positioning.positionerStyles],
+  );
+
+  const defaultProps: React.ComponentProps<'div'> = React.useMemo(() => {
+    const hiddenStyles: React.CSSProperties = {};
+
+    if (!open) {
+      hiddenStyles.pointerEvents = 'none';
+    }
+
+    return {
+      role: 'presentation',
+      hidden: !mounted,
+      style: {
+        ...positionerStyles,
+        ...hiddenStyles,
+      },
+    };
+  }, [open, mounted, positionerStyles]);
+
   const state: SelectPositioner.State = React.useMemo(
     () => ({
       open,
-      side: positioner.side,
-      align: positioner.align,
-      anchorHidden: positioner.anchorHidden,
+      side: positioning.side,
+      align: positioning.align,
+      anchorHidden: positioning.anchorHidden,
     }),
-    [open, positioner.side, positioner.align, positioner.anchorHidden],
+    [open, positioning.side, positioning.align, positioning.anchorHidden],
   );
 
   const setPositionerElement = useEventCallback((element) => {
@@ -134,7 +144,7 @@ export const SelectPositioner = React.forwardRef(function SelectPositioner(
     ref: [forwardedRef, setPositionerElement],
     state,
     customStyleHookMapping: popupStateMapping,
-    props: [positioner.getPositionerProps, elementProps],
+    props: [defaultProps, elementProps],
   });
 
   const prevMapSizeRef = React.useRef(0);
@@ -177,9 +187,18 @@ export const SelectPositioner = React.forwardRef(function SelectPositioner(
     }
   });
 
+  const contextValue: SelectPositionerContext = React.useMemo(
+    () => ({
+      ...positioning,
+      alignItemWithTriggerActive,
+      setControlledAlignItemWithTrigger,
+    }),
+    [positioning, alignItemWithTriggerActive, setControlledAlignItemWithTrigger],
+  );
+
   return (
     <CompositeList elementsRef={listRef} labelsRef={labelsRef} onMapChange={onMapChange}>
-      <SelectPositionerContext.Provider value={positioner}>
+      <SelectPositionerContext.Provider value={contextValue}>
         {mounted && modal && <InternalBackdrop inert={inertValue(!open)} />}
         {element}
       </SelectPositionerContext.Provider>
@@ -196,6 +215,14 @@ export namespace SelectPositioner {
   }
 
   export interface Props
-    extends useSelectPositioner.SharedParameters,
-      BaseUIComponentProps<'div', State> {}
+    extends useAnchorPositioning.SharedParameters,
+      BaseUIComponentProps<'div', State> {
+    /**
+     * Whether the positioner overlaps the trigger so the selected item's text is aligned
+     * with the trigger's value text. This only applies to mouse input and is automatically
+     * disabled if there is not enough space.
+     * @default true
+     */
+    alignItemWithTrigger?: boolean;
+  }
 }
