+++ packages/react/src/select/positioner/SelectPositionerContext.ts
@@ -1,13 +1,18 @@
 import * as React from 'react';
-import type { useSelectPositioner } from './useSelectPositioner';
+import { useAnchorPositioning } from '../../utils/useAnchorPositioning';
 
-export type SelectPositionerContext = useSelectPositioner.ReturnValue;
+export interface SelectPositionerContext extends useAnchorPositioning.ReturnValue {
+  alignItemWithTriggerActive: boolean;
+  setControlledAlignItemWithTrigger: React.Dispatch<React.SetStateAction<boolean>>;
+}
 
-export const SelectPositionerContext = React.createContext<SelectPositionerContext | null>(null);
+export const SelectPositionerContext = React.createContext<SelectPositionerContext | undefined>(
+  undefined,
+);
 
 export function useSelectPositionerContext() {
   const context = React.useContext(SelectPositionerContext);
-  if (context === null) {
+  if (!context) {
     throw new Error(
       'Base UI: SelectPositionerContext is missing. SelectPositioner parts must be placed within <Select.Positioner>.',
     );
